refactor: repackage, fix entity, clean code

+++ services/order/src/main/java/com/eCommerce/order/orderline/model/OrderLineRequest.java
@@ -7,5 +7,5 @@ public class OrderLineRequest {
     private Integer id;
     private Integer orderId;
     private Integer productId;
-    private double quantity;
+    private Integer quantity;
 }
